+++ apps/web/src/views/labs/layout.vue
@@ -43,13 +43,7 @@
           </template>
           <n-skeleton v-else width="15em" class="ml-5" />
         </div>
-        <global-description
-          v-if="description"
-          :description="description"
-          :show-button="showButton"
-          :show-full-content="showFullContent"
-          @toggle="toggle"
-        />
+        <global-description v-if="description" :description="description" />
       </template>
     </content-layout>
     <div v-else class="w-full flex-center">
@@ -60,7 +54,6 @@
 
 <script setup lang="ts">
 import type { TabPaneProps } from "naive-ui/es/tabs"
-import { useBoolean } from "@/composables"
 import { useRouterPush } from "@/composables/useRouterPush"
 import { LabRole } from "@/enum"
 import ContentLayout from "@/layouts/content-layout/index.vue"
@@ -196,10 +189,6 @@ const title = computed(() => {
 const description = computed(() => {
   return labInfo.value?.description ?? ""
 })
-
-const showButton = computed(() => description.value.length > 1000)
-
-const { bool: showFullContent, toggle } = useBoolean(!showButton.value)
 
 const { reloadPage } = useAppStore()
 async function handleNewMember() {